Issue #22: Adjust disk image name
crc expects the libvirt disk image to be named 'crc', but with the
libmachine go module update, it now defaults to naming it 'crc.disk'.
This commit specifies an explicit name rather than relying on the
default name to get the image filename crc expects.

This resolves https://github.com/code-ready/machine-driver-libvirt/issues/22